requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0

# Machine Learning and NLP
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.2
transformers>=4.20.0

# Google Gemini API
google-generativeai>=0.3.0

# Web Scraping (maintaining compatibility)
selenium>=4.0.0
beautifulsoup4>=4.10.0
requests>=2.28.0

# Web Interface
streamlit>=1.25.0
streamlit-chat>=0.1.1

# Utilities
python-dotenv>=0.19.0
pydantic>=1.10.0
tqdm>=4.64.0
loguru>=0.6.0

# Data Processing
nltk>=3.7
spacy>=3.4.0

# Optional: For advanced features
# gradio>=3.0.0  # Alternative UI framework
# chromadb>=0.4.0  # Alternative vector database
# langchain>=0.0.200  # For advanced RAG features

# Development and Testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0